Fix rendering of lines extending out of bounds
Previously the Point and Line types used usize for coordinates.

They now use i64 values, which are converted to usize values when they
are plotted (or are ignored if they are out of the image dimensions).

+++ src/point.rs
@@ -34,12 +34,12 @@ impl Display for Color {
 
 #[derive(Clone, Copy)]
 pub struct Point {
-    pub x: usize,
-    pub y: usize,
+    pub x: i64,
+    pub y: i64,
 }
 
 impl Point {
-    pub fn xy(x: usize, y: usize) -> Point {
+    pub fn xy(x: i64, y: i64) -> Point {
         Point { x: x, y: y }
     }
